requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
pydantic-settings
python-dotenv
aiofiles
tenacity
